Move DefaultLegacyReporting out of support package to break dependency cycle
